+++ src/createUseTrackedState.js
@@ -5,7 +5,7 @@ import {
   useReducer,
 } from 'react';
 
-import { useIsomorphicLayoutEffect } from './utils';
+import { useIsomorphicLayoutEffect, useAffectedDebugValue } from './utils';
 import {
   STATE_CONTEXT_PROPERTY,
   SUBSCRIBE_CONTEXT_PROPERTY,
@@ -77,6 +77,10 @@ export const createUseTrackedState = (context) => {
       const unsubscribe = subscribe(callback);
       return unsubscribe;
     }, [subscribe]);
+    if (process.env.NODE_ENV !== 'production') {
+      // eslint-disable-next-line react-hooks/rules-of-hooks
+      useAffectedDebugValue(state, affected);
+    }
     const proxyCache = useRef(new WeakMap()); // per-hook proxyCache
     return createDeepProxy(state, affected, proxyCache.current);
   };